# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS FinOps Analytics Platform API
  version: 1.0.0
  description: |
    The SaaS FinOps Analytics Platform provides multi-tenant cloud cost management,
    AI-powered database analytics, and subscription billing capabilities.

    ## Authentication

    The API supports two authentication methods:

    ### JWT Bearer Token
    Obtain a token via `POST /auth/login`. Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire after **15 minutes**. Use `POST /auth/refresh` to obtain a new one.

    ### API Key
    Generate an API key via `POST /auth/api-keys`. Include it in the `X-API-Key` header:
    ```
    X-API-Key: <your_api_key>
    ```

    ## Rate Limiting

    Rate limits are enforced per user per minute based on subscription tier:

    | Plan       | Requests/min |
    |------------|-------------|
    | Free       | 100         |
    | Base       | 500         |
    | Pro        | 2,000       |
    | Enterprise | 10,000      |

    When the limit is exceeded, the API returns `429 Too Many Requests` with a `Retry-After` header.

  contact:
    name: Platform Support
    email: support@finops-platform.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: http://localhost:8081
    description: Auth Service (direct)
  - url: http://localhost:8082
    description: Billing Service (direct)
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login. Expires in 15 minutes.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated via POST /auth/api-keys.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid credentials"
        message:
          type: string
          example: "The provided email or password is incorrect"
      required: [error]
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "rate limit exceeded"
        retry_after:
          type: integer
          example: 30
          description: Seconds until the rate limit resets
...
    QueryExecuteRequest:
      type: object
      required: [database_connection_id, query_text]
      properties:
        database_connection_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        query_text:
          type: string
          example: "Show me top 10 customers by revenue this month"
    QueryExecuteResponse:
      type: object
      properties:
        chart_type:
          type: string
          enum: [bar, line, pie, table, metric]
          example: "bar"
        labels:
          type: array
          items:
            type: string
          example: ["Customer A", "Customer B", "Customer C"]
        data:
          type: array
          items:
            type: number
          example: [15000, 12000, 9500]
        raw_data:
          type: array
          items:
            type: object
          description: Full result rows as returned by the database
        generated_sql:
          type: string
          example: "SELECT customer_name, SUM(revenue) AS total_revenue FROM orders WHERE DATE_TRUNC('month', created_at) = DATE_TRUNC('month', NOW()) GROUP BY customer_name ORDER BY total_revenue DESC LIMIT 10"
        execution_time_ms:
          type: integer
          example: 245
        cached:
          type: boolean
          example: false
    QueryHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        query_text:
          type: string
        generated_sql:
          type: string
          nullable: true
        execution_time_ms:
          type: integer
          nullable: true
        result_count:
          type: integer
          nullable: true
        status:
          type: string
          enum: [success, error, timeout]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /query/execute:
    post:
      tags: [query]
      summary: Execute a natural language query
      description: |
        Converts a natural language query to SQL and executes it against the specified
        database. Results are automatically formatted for chart rendering.

        **Processing flow:**
        1. Fetch database schema
        2. Generate SQL using LLM (OpenAI/LangChain)
        3. Check Redis cache (5-minute TTL)
        4. Execute SQL against external database (30-second timeout)
        5. Determine chart type and format response

        **Chart types returned:**
        - `metric` - Single numeric value
        - `line` - Time series data
        - `bar` - Category comparison
        - `pie` - Part-to-whole relationships
        - `table` - Tabular/multi-column data

        **Rate limit:** 100 req/min (Free), 500 (Base), 2000 (Pro), 10000 (Enterprise)

        **Cached results** are returned within 100ms. Cache key = query_text + connection_id.

      operationId: executeQuery
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryExecuteRequest'
            example:
              database_connection_id: "550e8400-e29b-41d4-a716-446655440002"
              query_text: "Show me top 10 customers by revenue this month"
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryExecuteResponse'
              example:
                chart_type: "bar"
                labels: ["Acme Corp", "Globex", "Initech", "Umbrella", "Cyberdyne"]
                data: [152000, 98500, 87200, 76400, 65100]
                raw_data:
                  - customer_name: "Acme Corp"
                    total_revenue: 152000
                  - customer_name: "Globex"
                    total_revenue: 98500
                generated_sql: "SELECT customer_name, SUM(amount) AS total_revenue FROM orders JOIN customers ON orders.customer_id = customers.id WHERE DATE_TRUNC('month', orders.created_at) = DATE_TRUNC('month', NOW()) GROUP BY customer_name ORDER BY total_revenue DESC LIMIT 10"
                execution_time_ms: 245
                cached: false
        '400':
          description: SQL generation failed or invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "could not generate SQL: query is ambiguous - please specify which table to use"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Database connection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '408':
          description: Query execution timed out (30-second limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "query execution timed out after 30 seconds. Try a more specific query."
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-code-samples:
        - lang: Python
          label: Python (requests)
          source: |
            import requests

            API_BASE = "http://localhost:8084"
            TOKEN = "your_access_token_here"

            response = requests.post(
                f"{API_BASE}/query/execute",
                headers={
                    "Authorization": f"Bearer {TOKEN}",
                    "Content-Type": "application/json",
                },
                json={
                    "database_connection_id": "550e8400-e29b-41d4-a716-446655440002",
                    "query_text": "Show me top 10 customers by revenue this month",
                },
            )

            result = response.json()
            print(f"Chart type: {result['chart_type']}")
            print(f"Generated SQL: {result['generated_sql']}")
            print(f"Execution time: {result['execution_time_ms']}ms")

        - lang: JavaScript
          label: JavaScript (fetch)
          source: |
            const API_BASE = 'http://localhost:8084';
            const TOKEN = 'your_access_token_here';

            const response = await fetch(`${API_BASE}/query/execute`, {
              method: 'POST',
              headers: {
                'Authorization': `Bearer ${TOKEN}`,
                'Content-Type': 'application/json',
              },
              body: JSON.stringify({
                database_connection_id: '550e8400-e29b-41d4-a716-446655440002',
                query_text: 'Show me top 10 customers by revenue this month',
              }),
            });

            const result = await response.json();
            console.log('Chart type:', result.chart_type);
            console.log('Generated SQL:', result.generated_sql);
            console.log('Execution time:', result.execution_time_ms, 'ms');

        - lang: Go
          label: Go (net/http)
          source: |
            package main

            import (
                "bytes"
                "encoding/json"
                "fmt"
                "net/http"
            )

            func main() {
                apiBase := "http://localhost:8084"
                token := "your_access_token_here"

                payload := map[string]string{
                    "database_connection_id": "550e8400-e29b-41d4-a716-446655440002",
                    "query_text":             "Show me top 10 customers by revenue this month",
                }
                body, _ := json.Marshal(payload)

                req, _ := http.NewRequest("POST", apiBase+"/query/execute", bytes.NewBuffer(body))
                req.Header.Set("Authorization", "Bearer "+token)
                req.Header.Set("Content-Type", "application/json")

                client := &http.Client{}
                resp, err := client.Do(req)
                if err != nil {
                    panic(err)
                }
                defer resp.Body.Close()

                var result map[string]interface{}
                json.NewDecoder(resp.Body).Decode(&result)
                fmt.Printf("Chart type: %v\n", result["chart_type"])
                fmt.Printf("Generated SQL: %v\n", result["generated_sql"])
            }
  /query/history:
    get:
      tags: [query]
      summary: Get query history
      description: Returns the query execution history for the authenticated user.
      operationId: getQueryHistory
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: connection_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by database connection
      responses:
        '200':
          description: Query history
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryHistoryItem'
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─────────────────────────────────────────────
  # Health check (all services)
  # ─────────────────────────────────────────────
...